Move deprecated methods to separate deprecated files
